Доработал модуль mocks, удалил лишнее
