--- js/data.js
@@ -1,8 +1,0 @@
-// Массив имен
-const NAMES = ['Казимир', 'Карл', 'Карен', 'Клавдий', 'Кирилл', 'Клим', 'Ким', 'Клод', 'Клемент', 'Корнилий', 'Корней', 'Конрад', 'Конкордий', 'Константин', 'Кондрат', 'Ксаннф', 'Кузьма'];
-
-//Массив комментариев
-const COMMENTS = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
-
-
-export { NAMES, COMMENTS };
