+++ js/form.js
@@ -1,6 +1,6 @@
-import { isEscapeKey } from './utils.js';
+import { isEscapeKey, checkingStringLength } from './utils.js';
 import { REGEX, MAX_HASHTAGS, MAX_LENGTH_DESCRIPTION } from './setup.js';
-import { checkingStringLength } from './utils.js';
+
 
 const uploadInput = document.querySelector('#upload-file');
 const editeUploadImg = document.querySelector('.img-upload__overlay');
