+++ js/main.js
@@ -2,7 +2,7 @@
 import { getPreviewPhoto } from './preview.js';
 import { onPictureClick } from './fullscreen.js';
 import { onUploadImg } from './form.js';
-import { createDescription } from './moks.js';
+import { createDescription } from './mocks.js';
 
 const arrayObjects = Array.from({ length: 25 }, (item, i) => createDescription(i));
 getPreviewPhoto();
